+++ Nbm/src/test/resources/projForTest/src/indentation/pasteAtBlankLineInBody.kt
@@ -1,0 +1,5 @@
+package indentation
+
+fun foo() {
+<caret>
+}