+++ Nbm/src/test/resources/projForTest/src/indentation/pasteWithOwnIndent.kt
@@ -1,0 +1,5 @@
+package indentation
+
+fun qux() {
+<caret>
+}